style(ui): move ConnectionSelector dropdown above query editor
