+++ src/renderer/src/components/workspace/query-panel/connection-selector.tsx
